Move Moveable into it's own file
- Also moving San into its own file
- Simplify Uci and San trait
