+++ core/src/main/scala/Moveable.scala
@@ -1,0 +1,5 @@
+package chess
+
+trait Moveable:
+  def apply(position: Position): Either[ErrorStr, MoveOrDrop]
+  def rawString: Option[String]
